Fix calculation of stock indicators

+++ build_project.sh
@@ -1,3 +1,9 @@
+if [ $# -eq 0 ]
+  then
+    echo "No mail_password provided"
+    exit 1
+fi
+
 perl -pi -e 's/192.168.99.100/127.0.0.1/g' configs/rabbitmq/properties/rabbitmq_config.properties
 perl -pi -e 's/mail_password/'$1'/g' web_app/gpw_radar/src/main/resources/config/application.yml
 rm -rf target
